Requirement: #1709 - Serialization to DICOM GSPS Requirement: #1710 - Deserialization from DICOM GSPS * Changed PixelAspectRatio to use ints instead of doubles (as they should be) * Modified DisplayedAreaModuleIod to use common structures such as Point and Size instead of returning int and float arrays. * Fixed names of ImageHorizontalFlip enumerated values * Added ImageSopInstanceReferenceDictionary and SeriesReferenceDictionary as utilities for testing presence of a reference to a given series/instance/frame/segment combination * Added ConvertToPixelAddressRectangle helper to RectangleUtilities * Added base classes for serialization and deserialization of xSPS objects * Added classes for serialization and deserialization of GSPS objects * Added a debug tool for manually saving and loading xSPS objects.

+++ ImageViewer/StudyManagement/Frame.cs
@@ -532,8 +532,8 @@ namespace ClearCanvas.ImageViewer.StudyManagement
 				pixelAspectRatio = _parentImageSop[DicomTags.PixelAspectRatio].ToString();
 				if (!string.IsNullOrEmpty(pixelAspectRatio))
 				{
-					double[] values;
-					if (DicomStringHelper.TryGetDoubleArray(pixelAspectRatio, out values) && values.Length == 2)
+					int[] values;
+					if (DicomStringHelper.TryGetIntArray(pixelAspectRatio, out values) && values.Length == 2)
 						return new PixelAspectRatio(values[0], values[1]);
 				}
